Centralize use retry fetch to util, remove duplicate passthrough funcs
